feat: add `longFieldMaxLength` config option, deprecate `errorMessageMaxLength` (#2193)
This config option will truncate some transaction, span, and error fields
that can be longer at the configured number of JavaScript characters per
https://github.com/elastic/apm/blob/master/specs/agents/field-limits.md#long_field_max_length-configuration

Notably truncation is now based on a count of JavaScript *characters*
rather than in UTF-8 encoded bytes.

This also deprecates the errorMessageMaxLength config option. It now
defaults to the longFieldMaxLength value if not specified. Otherwise
errorMessageMaxLength is still supported.

Fixes: #1921